Anpassen der delete Item funktion - prüft ob item noch ausgeliehen ist und warnt manager. Soll dennoch gelöxscht werden werden aktive loans beendet

+++ Frontend/BAMBUS/src/store/modules/item/actions.js
@@ -2,6 +2,7 @@ import ItemServices from "../../services/ItemServices";
 import store from "../../index.js";
 import $router from "@/router";
 import item from ".";
+import LoanService from "@/store/services/LoanService";
 
 export default {
   async loadItems({ commit }) {
@@ -14,6 +15,20 @@ export default {
     });
   },
   async deleteItem({ commit}, id) {
+    let item = store.getters["itemStore/getItemById"](id);
+    console.log(item.currentLoanId);
+    if (item.currentLoanId != 0) {
+      if (confirm("Dieser Artikel ist noch ausgeliehen. Wollen Sie ihn trotzdem löschen?") == false) {
+        console.log("DOnt delete item")
+        return;
+      }
+      else {
+        console.log("Set return date and end extension request");
+        LoanService.SetReturnDate(item.currentLoanId);
+        LoanService.EndExtensionRequest(item.currentLoanId);
+      }
+    }
+    console.log("Delete item"); 
     await ItemServices.DeleteItem(id).then((response) => {
       if (response.data.success) {
         commit("setItems", response.data.data);
